requirements.txt
APScheduler==3.0.1
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
cssselect==0.9.1
futures==2.2.0
itsdangerous==0.24
lxml==3.4.1
pytz==2014.10
requests==2.4.3
six==1.8.0
tzlocal==1.1.2
wsgiref==0.1.2
python-dateutil==2.2